More HTTP endpoints separated
