+++ app.js
@@ -1,11 +1,6 @@
 // Imported modules
 const express = require('express')
 const session = require('express-session') // For storing client login data
-const { encrypt, decrypt } = require('./crypto.js') // For encrypting passwords
-const jwt = require('jsonwebtoken') // For authentication system between Plugins and Formbar
-const excelToJson = require('convert-excel-to-json')
-const multer = require('multer') // Used to upload files
-const upload = multer({ dest: 'uploads/' }) // Selects a file destination for uploaded files to go to, will create folder when file is submitted(?)
 const crypto = require('crypto')
 const fs = require("fs")
 const { isAuthenticated, isLoggedIn, permCheck } = require('./modules/authentication.js')
@@ -13,12 +8,11 @@ const { logger } = require('./modules/logger.js')
 const { logNumbers, settings } = require('./modules/config.js')
 const { MANAGER_PERMISSIONS, TEACHER_PERMISSIONS, GUEST_PERMISSIONS, STUDENT_PERMISSIONS, MOD_PERMISSIONS, BANNED_PERMISSIONS, DEFAULT_CLASS_PERMISSIONS, CLASS_SOCKET_PERMISSIONS, CLASS_SOCKET_PERMISSION_SETTINGS, GLOBAL_SOCKET_PERMISSIONS } = require('./modules/permissions.js')
 const { classInformation, Classroom } = require('./modules/class.js')
-const { Student } = require('./modules/student.js')
-const database = require('./modules/database.js')
+const { database } = require('./modules/database.js')
+const { initSocketRoutes, advancedEmitToClass, managerUpdate, io } = require("./sockets/init.js")
 
 const app = express()
 const http = require('http').createServer(app)
-const io = require('socket.io')(http)
 
 // Set EJS as our view engine
 app.set('view engine', 'ejs')
@@ -368,7 +362,7 @@ async function getIpAccess(type) {
 	}, {})
 }
 
-// Import routes
+// Import HTTP routes
 const apiRoutes = require('./routes/api.js')(classInformation)
 const routeFiles = fs.readdirSync('./routes/').filter(file => file.endsWith('.js'));
 
@@ -384,181 +378,6 @@ for (const routeFile of routeFiles) {
 
 // Add routes to express
 app.use('/api', apiRoutes)
-
-// An endpoint for the teacher to control the formbar
-// Used to update students permissions, handle polls and their corresponsing responses
-// On render it will send all students in that class to the page
-app.get('/controlPanel', isAuthenticated, permCheck, (req, res) => {
-	try {
-		logger.log('info', `[get /controlPanel] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-
-		let students = classInformation[req.session.class].students
-		let keys = Object.keys(students)
-		let allStuds = []
-
-		for (var i = 0; i < keys.length; i++) {
-			var val = { name: keys[i], perms: students[keys[i]].permissions, pollRes: { lettRes: students[keys[i]].pollRes.buttonRes, textRes: students[keys[i]].pollRes.textRes }, help: students[keys[i]].help }
-			allStuds.push(val)
-		}
-
-		/* Uses EJS to render the template and display the information for the class.
-		This includes the class list of students, poll responses, and the class code - Riley R., May 22, 2023
-		*/
-		res.render('pages/controlPanel', {
-			title: 'Control Panel',
-			pollStatus: classInformation[req.session.class].poll.status,
-			settingsPermissions: classInformation[req.session.class].permissions.manageClass,
-			tagNames: classInformation[req.session.class].tagNames
-		})
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-// C
-
-/*
-Manages the use of excell spreadsheets in order to create progressive lessons.
-It uses Excel To JSON to create an object containing all the data needed for a progressive lesson.
-Could use a switch if need be, but for now it's all broken up by if statements.
-Use the provided template when testing things. - Riley R., May 22, 2023
-*/
-app.post('/controlPanel', upload.single('spreadsheet'), isAuthenticated, permCheck, (req, res) => {
-	try {
-		//Initialze a list to push each step to - Riley R., May 22, 2023
-		let steps = []
-
-		logger.log('info', `[post /controlPanel] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-
-		/*
-		Uses Excel to JSON to read the sent excel spreadsheet.
-		Each main column has been assigned a label in order to differentiate them.
-		It loops through the whole object - Riley R., May 22, 2023
-		*/
-		if (req.file) {
-			classInformation[req.session.class].currentStep = 0
-			const result = excelToJson({
-				sourceFile: req.file.path,
-				sheets: [{
-					name: 'Steps',
-					columnToKey: {
-						A: 'index',
-						B: 'type',
-						C: 'prompt',
-						D: 'response',
-						E: 'labels'
-					}
-				}]
-			})
-
-			/* For In Loop that iterates through the created object.
-			 Allows for the use of steps inside of a progressive lesson.
-			 Checks the object's type using a conditional - Riley R., May 22, 2023
-			*/
-			for (const key in result['Steps']) {
-				let step = {}
-				// Creates an object with all the data required to start a poll - Riley R., May 22, 2023
-				if (result['Steps'][key].type == 'Poll') {
-					step.type = 'poll'
-					step.labels = result['Steps'][key].labels.split(', ')
-					step.responses = result['Steps'][key].response
-					step.prompt = result['Steps'][key].prompt
-					steps.push(step)
-					// Creates an object with all the data required to start a quiz
-				} else if (result['Steps'][key].type == 'Quiz') {
-					let nameQ = result['Steps'][key].prompt
-					let letters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
-					let colToKeyObj = {
-						A: 'index',
-						B: 'question',
-						C: 'key'
-					}
-					let i = 0
-					/*
-					Names the cells of the sheet after C to A-Z for the use of them in Quizzes (A, B, and C in the spreadsheet are the index, question, and key, not the answers)
-					Creates a way to have multiple responses to quizzes- Riley R., May 22, 2023
-					*/
-					for (const letterI in letters) {
-						if (letters.charAt(letterI) != 'A' && letters.charAt(letterI) != 'B' && letters.charAt(letterI) != 'C') {
-							colToKeyObj[letters.charAt(letterI)] = letters.charAt(i)
-							i++
-						}
-					}
-					let quizLoad = excelToJson({
-						sourceFile: req.file.path,
-						sheets: [{
-							name: nameQ,
-							columnToKey: colToKeyObj
-						}]
-					})
-					let questionList = []
-					for (let i = 1; i < quizLoad[nameQ].length; i++) {
-						let questionMaker = []
-
-						questionMaker.push(quizLoad[nameQ][i].question)
-						questionMaker.push(quizLoad[nameQ][i].key)
-						for (const letterI in letters) {
-							if (quizLoad[nameQ][i][letters.charAt(letterI)] != undefined) {
-								questionMaker.push(quizLoad[nameQ][i][letters.charAt(letterI)])
-							}
-						}
-						questionList.push(questionMaker)
-					}
-					step.type = 'quiz'
-					step.questions = questionList
-					steps.push(step)
-				} else if (result['Steps'][key].type == 'Lesson') {
-					/*
-					Creates an object with all necessary data in order to make a lesson.
-					The data is stored on a page in an excel spreadsheet.
-					the name of this page is defined in the main page of the excel spreadsheet. - Riley R., May 22, 2023
-					*/
-					nameL = result['Steps'][key].prompt
-					let lessonLoad = excelToJson({
-						sourceFile: req.file.path,
-						sheets: [{
-							name: nameL,
-							columnToKey: {
-								A: 'header',
-								B: 'data'
-							}
-						}]
-					})
-					let lessonArr = []
-					for (let i = 1; i < lessonLoad[nameL].length; i++) {
-						let lessonMaker = [lessonLoad[nameL][i].header]
-
-						let lessonContent = lessonLoad[nameL][i].data.split(', ')
-						for (let u = 0; u < lessonContent.length; u++) {
-							lessonMaker.push(lessonContent[u])
-						}
-						lessonArr.push(lessonMaker)
-					}
-
-					let dateConfig = new Date()
-
-					step.type = 'lesson'
-					step.date = `${dateConfig.getMonth() + 1}/${dateConfig.getDate()}/${dateConfig.getFullYear()}`
-					step.lesson = lessonArr
-					steps.push(step)
-				}
-			}
-
-			classInformation[req.session.class].steps = steps
-			res.redirect('/controlPanel')
-		}
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
 
 // Allow teacher to create class
 // Allowing the teacher to create classes is vital to whether the lesson actually works or not, because they have to be allowed to create a teacher class
@@ -576,7 +395,7 @@ app.post('/createClass', isLoggedIn, permCheck, (req, res) => {
 		async function makeClass(id, className, key, permissions, sharedPolls = [], pollHistory = [], tags) {
 			try {
 				// Get the teachers session data ready to transport into new class
-				var user = classInformation.noClass.students[req.session.username]
+				const user = classInformation.noClass.students[req.session.username]
 
 				logger.log('verbose', `[makeClass] id=(${id}) name=(${className}) key=(${key}) sharedPolls=(${JSON.stringify(sharedPolls)})`)
 				// Remove teacher from no class
@@ -734,258 +553,6 @@ app.post('/createClass', isLoggedIn, permCheck, (req, res) => {
 	}
 })
 
-// L
-// This renders the login page
-// It displays the title and the color of the login page of the formbar js
-// It allows for the login to check if the user wants to login to the server
-// This makes sure the lesson can see the students and work with them
-app.get('/login', (req, res) => {
-	try {
-		logger.log('info', `[get /login] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-
-		res.render('pages/login', {
-			title: 'Login'
-		})
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-// This lets the user log into the server, it uses each element from the database to allow the server to do so
-// This lets users actually log in instead of not being able to log in at all
-// It uses the usernames, passwords, etc. to verify that it is the user that wants to log in logging in
-// This also encrypts passwords to make sure people's accounts don't get hacked
-app.post('/login', async (req, res) => {
-	try {
-		var user = {
-			username: req.body.username,
-			password: req.body.password,
-			loginType: req.body.loginType,
-			userType: req.body.userType,
-			displayName: req.body.displayName
-		}
-		var passwordCrypt = encrypt(user.password)
-
-		logger.log('info', `[post /login] ip=(${req.ip}) session=(${JSON.stringify(req.session)}`)
-		logger.log('verbose', `[post /login] username=(${user.username}) password=(${Boolean(user.password)}) loginType=(${user.loginType}) userType=(${user.userType})`)
-
-		// Check whether user is logging in or signing up
-		if (user.loginType == 'login') {
-			logger.log('verbose', '[post /login] User is logging in')
-
-			// Get the users login in data to verify password
-			database.get('SELECT users.*, CASE WHEN shared_polls.pollId IS NULL THEN json_array() ELSE json_group_array(DISTINCT shared_polls.pollId) END as sharedPolls, CASE WHEN custom_polls.id IS NULL THEN json_array() ELSE json_group_array(DISTINCT custom_polls.id) END as ownedPolls FROM users LEFT JOIN shared_polls ON shared_polls.userId = users.id LEFT JOIN custom_polls ON custom_polls.owner = users.id WHERE users.username=?', [user.username], async (err, userData) => {
-				try {
-					// Check if a user with that name was not found in the database
-					if (!userData.username) {
-						logger.log('verbose', '[post /login] User does not exist')
-						res.render('pages/message', {
-							message: 'No user found with that username.',
-							title: 'Login'
-						})
-						return
-					}
-
-					if (!userData.displayName) {
-						database.run("UPDATE users SET displayName = ? WHERE username = ?", [userData.username, userData.username]), (err) => {
-							try {
-								if (err) throw err;
-								logger.log('verbose', '[post /login] Added displayName to database');
-							} catch (err) {
-								logger.log('error', err.stack);
-								res.render('pages/message', {
-									message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-									title: 'Error'
-								});
-							};
-						};
-					};
-
-					// Decrypt users password
-					let tempPassword = decrypt(JSON.parse(userData.password))
-					if (tempPassword != user.password) {
-						logger.log('verbose', '[post /login] Incorrect password')
-						res.render('pages/message', {
-							message: 'Incorrect password',
-							title: 'Login'
-						})
-						return
-					}
-
-					let loggedIn = false
-					let classKey = ''
-
-					for (let classData of Object.values(classInformation)) {
-						if (classData.key) {
-							for (let username of Object.keys(classData.students)) {
-								if (username == userData.username) {
-									loggedIn = true
-									classKey = classData.key
-
-									break
-								}
-							}
-						}
-					}
-
-					if (loggedIn) {
-						logger.log('verbose', '[post /login] User is already logged in')
-						req.session.class = classKey
-					} else {
-						// Add user to the session
-						classInformation.noClass.students[userData.username] = new Student(
-							userData.username,
-							userData.id,
-							userData.permissions,
-							userData.API,
-							JSON.parse(userData.ownedPolls),
-							JSON.parse(userData.sharedPolls),
-							userData.tags,
-							userData.displayName
-
-						)
-						req.session.class = 'noClass'
-					}
-					// Add a cookie to transfer user credentials across site
-					req.session.userId = userData.id
-					req.session.username = userData.username
-					req.session.tags = userData.tags
-					req.session.displayName = userData.displayName
-
-					logger.log('verbose', `[post /login] session=(${JSON.stringify(req.session)})`)
-					logger.log('verbose', `[post /login] cD=(${JSON.stringify(classInformation)})`)
-
-					res.redirect('/')
-				} catch (err) {
-					logger.log('error', err.stack);
-					res.render('pages/message', {
-						message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-						title: 'Error'
-					})
-				}
-			})
-		} else if (user.loginType == 'new') {
-			logger.log('verbose', '[post /login] Creating new user')
-
-			let permissions = STUDENT_PERMISSIONS
-
-			database.all('SELECT API, secret, username FROM users', (err, users) => {
-				try {
-					if (err) throw err
-
-					let existingAPIs = []
-					let existingSecrets = []
-					let newAPI
-					let newSecret
-
-					if (users.length == 0) permissions = MANAGER_PERMISSIONS
-
-					for (let dbUser of users) {
-						existingAPIs.push(dbUser.API)
-						existingSecrets.push(dbUser.secret)
-						if (dbUser.username == user.username) {
-							logger.log('verbose', '[post /login] User already exists')
-							res.render('pages/message', {
-								message: 'A user with that username already exists.',
-								title: 'Login'
-							})
-							return
-						}
-					}
-
-					do {
-						newAPI = crypto.randomBytes(64).toString('hex')
-					} while (existingAPIs.includes(newAPI))
-					do {
-						newSecret = crypto.randomBytes(256).toString('hex')
-					} while (existingSecrets.includes(newSecret))
-
-					// Add the new user to the database
-					database.run(
-						'INSERT INTO users(username, password, permissions, API, secret, displayName) VALUES(?, ?, ?, ?, ?, ?)',
-						[
-							user.username,
-							JSON.stringify(passwordCrypt),
-							permissions,
-							newAPI,
-							newSecret,
-							user.displayName
-						], (err) => {
-							try {
-								if (err) throw err
-
-								logger.log('verbose', '[post /login] Added user to database')
-
-								// Find the user in which was just created to get the id of the user
-								database.get('SELECT * FROM users WHERE username=?', [user.username], (err, userData) => {
-									try {
-										if (err) throw err
-
-										// Add user to session
-										classInformation.noClass.students[userData.username] = new Student(
-											userData.username,
-											userData.id,
-											userData.permissions,
-											userData.API,
-											[],
-											[],
-											userData.tags,
-											userData.displayName
-										)
-
-										// Add the user to the session in order to transfer data between each page
-										req.session.userId = userData.id
-										req.session.username = userData.username
-										req.session.class = 'noClass'
-										req.session.displayName = userData.displayName;
-
-										logger.log('verbose', `[post /login] session=(${JSON.stringify(req.session)})`)
-										logger.log('verbose', `[post /login] cD=(${JSON.stringify(classInformation)})`)
-
-										managerUpdate()
-
-										res.redirect('/')
-									} catch (err) {
-										logger.log('error', err.stack);
-										res.render('pages/message', {
-											message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-											title: 'Error'
-										})
-									}
-								})
-							} catch (err) {
-								logger.log('error', err.stack);
-								res.render('pages/message', {
-									message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-									title: 'Error'
-								})
-							}
-						}
-					)
-				} catch (err) {
-					logger.log('error', err.stack);
-					res.render('pages/message', {
-						message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-						title: 'Error'
-					})
-				}
-			})
-		} else if (user.loginType == 'guest') {
-			logger.log('verbose', '[post /login] Logging in as guest')
-		}
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
 // M
 // Loads which classes the teacher is an owner of
 // This allows the teacher to be in charge of all classes
@@ -1015,118 +582,6 @@ app.get('/managerPanel', isLoggedIn, permCheck, (req, res) => {
 
 		res.render('pages/managerPanel', {
 			title: 'Manager Panel'
-		})
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-// N
-
-// O
-/* This is what happens when the server tries to authenticate a user. It saves the redirectURL query parameter to a variable, and sends the redirectURL to the oauth page as
-a variable. */
-app.get('/oauth', (req, res) => {
-	try {
-		let redirectURL = req.query.redirectURL
-
-		logger.log('info', `[get /oauth] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-		logger.log('verbose', `[get /oauth] redirectURL=(${redirectURL})`)
-
-		res.render('pages/oauth.ejs', {
-			title: 'Oauth',
-			redirectURL: redirectURL
-		})
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-// This is what happens after the user submits their authentication data.
-app.post('/oauth', (req, res) => {
-	try {
-		// It saves the username, password, and the redirectURL that is submitted.
-		const {
-			username,
-			password,
-			redirectURL
-		} = req.body
-
-		logger.log('info', `[post /oauth] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-		logger.log('verbose', `[post /oauth] username=(${username}) redirectURL=(${redirectURL})`)
-
-		if (!username) {
-			res.render('pages/message', {
-				message: 'Please enter a username',
-				title: 'Login'
-			})
-			return
-		}
-		if (!password) {
-			res.render('pages/message', {
-				message: 'Please enter a password',
-				title: 'Login'
-			})
-			return
-		}
-
-		database.get('SELECT * FROM users WHERE username=?', [username], (err, userData) => {
-			try {
-				if (err) throw err
-
-				// Check if a user with that name was not found in the database
-				if (!userData.username) {
-					logger.log('verbose', '[post /oauth] User does not exist')
-					res.render('pages/message', {
-						message: 'No user found with that username.',
-						title: 'Login'
-					})
-					return
-				}
-
-				// Decrypt users password
-				let databasePassword = decrypt(JSON.parse(userData.password))
-				if (databasePassword != password) {
-					logger.log('verbose', '[post /oauth] Incorrect password')
-					res.render('pages/message', {
-						message: 'Incorrect password',
-						title: 'Login'
-					})
-					return
-				}
-
-				let classCode = getUserClass(userData.username)
-
-				userData.classPermissions = null
-
-				if (classInformation[classCode] && classInformation[classCode].students[userData.username])
-					userData.classPermissions = classInformation[classCode].students[userData.username].classPermissions
-
-				var token = jwt.sign({
-					id: userData.id,
-					username: userData.username,
-					permissions: userData.permissions,
-					classPermissions: userData.classPermissions,
-					class: classCode
-				}, userData.secret, { expiresIn: '30m' })
-
-				logger.log('verbose', '[post /oauth] Successfully Logged in with oauth')
-				res.redirect(`${redirectURL}?token=${token}`)
-			} catch (err) {
-				logger.log('error', err.stack);
-				res.render('pages/message', {
-					message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-					title: 'Error'
-				})
-			}
 		})
 	} catch (err) {
 		logger.log('error', err.stack);
@@ -1189,236 +644,6 @@ app.post('/previousLessons', isAuthenticated, permCheck, (req, res) => {
 	}
 })
 
-app.get('/plugins', isAuthenticated, permCheck, (req, res) => {
-	try {
-		logger.log('info', `[get /plugins] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-
-		res.render('pages/plugins.ejs', {
-			title: 'Plugins'
-		})
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-// Q
-
-
-
-// R
-
-
-// S
-app.get('/selectClass', isLoggedIn, permCheck, (req, res) => {
-	try {
-		logger.log('info', `[get /selectClass] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-
-		database.all(
-			'SELECT classroom.name, classroom.key FROM users JOIN classusers ON users.id = classusers.studentId JOIN classroom ON classusers.classId = classroom.id WHERE users.username=?',
-			[req.session.username],
-			(err, joinedClasses) => {
-				try {
-					if (err) throw err
-
-					logger.log('verbose', `[get /selectClass] joinedClasses=(${JSON.stringify(joinedClasses)})`)
-					res.render('pages/selectClass', {
-						title: 'Select Class',
-						joinedClasses: joinedClasses
-					})
-				} catch (err) {
-					logger.log('error', err.stack);
-					res.render('pages/message', {
-						message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-						title: 'Error'
-					})
-				}
-			}
-		)
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-
-//Adds user to a selected class, typically from the select class page
-app.post('/selectClass', isLoggedIn, permCheck, async (req, res) => {
-	try {
-		let classCode = req.body.key.toLowerCase()
-
-		logger.log('info', `[post /selectClass] ip=(${req.ip}) session=(${JSON.stringify(req.session)}) classCode=(${classCode})`)
-
-		let classJoinStatus = await joinClass(req.session.username, classCode)
-
-		if (typeof classJoinStatus == 'string') {
-			res.render('pages/message', {
-				message: `Error: ${classJoinStatus}`,
-				title: 'Error'
-			})
-			return
-		}
-
-		let classData = classInformation[classCode]
-
-		let cpPermissions = Math.min(
-			classData.permissions.controlPolls,
-			classData.permissions.manageStudents,
-			classData.permissions.manageClass
-		)
-
-		advancedEmitToClass('cpUpdate', classCode, { classPermissions: cpPermissions }, classInformation[classCode])
-
-		req.session.class = classCode
-
-		setClassOfApiSockets(classInformation[classCode].students[req.session.username].API, classCode)
-
-		res.redirect('/')
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-
-/* Student page, the layout is controlled by different "modes" to display different information.
-There are currently 3 working modes
-Poll: For displaying a multiple choice or essay question
-Quiz: Displaying a quiz with questions that can be answered by the student
-Lesson: used to display an agenda of sorts to the stufent, but really any important info can be put in a lesson - Riley R., May 22, 2023
-*/
-app.get('/student', isAuthenticated, permCheck, (req, res) => {
-	try {
-		//Poll Setup
-		let user = {
-			name: req.session.username,
-			class: req.session.class,
-			tags: req.session.tags
-		}
-		let answer = req.query.letter
-
-		logger.log('info', `[get /student] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-		logger.log('verbose', `[get /student] question=(${JSON.stringify(req.query.question)}) answer=(${req.query.letter})`)
-
-		if (answer) {
-			classInformation[req.session.class].students[req.session.username].pollRes.buttonRes = answer
-		}
-
-		//Quiz Setup and Queries
-		/* Sets up the query parameters you can enter when on the student page. These return either a question by it's index or a question by a randomly generated index.
-
-		formbar.com/students?question=random or formbar.com/students?question=[number] are the params you can enter at the current moment.
-
-		If you did not enter a query the page will be loaded normally. - Riley R., May 24, 2023
-		*/
-		if (req.query.question == 'random') {
-			let random = Math.floor(Math.random() * classInformation[req.session.class].quiz.questions.length)
-
-			logger.log('verbose', `[get /student] quiz=(${JSON.stringify(classInformation[req.session.class].quiz.questions[random])})`)
-
-			res.render('pages/queryquiz', {
-				quiz: JSON.stringify(classInformation[req.session.class].quiz.questions[random]),
-				title: 'Quiz'
-			})
-			if (classInformation[req.session.class].quiz.questions[req.query.question] != undefined) {
-				logger.log('verbose', `[get /student] quiz=(${JSON.stringify(classInformation[req.session.class].quiz.questions[req.query.question])})`)
-
-				res.render('pages/queryquiz', {
-					quiz: JSON.stringify(classInformation[req.session.class].quiz.questions[random]),
-					title: 'Quiz'
-				})
-			}
-		} else if (isNaN(req.query.question) == false) {
-			if (typeof classInformation[req.session.class].quiz.questions[req.query.question] != 'undefined') {
-				logger.log('verbose', `[get /student] quiz=(${JSON.stringify(classInformation[req.session.class].quiz.questions[req.query.question])})`)
-
-				res.render('pages/queryquiz', {
-					quiz: JSON.stringify(classInformation[req.session.class].quiz.questions[req.query.question]),
-					title: 'Quiz'
-				})
-			} else {
-				res.render('pages/message', {
-					message: 'Error: please enter proper data',
-					title: 'Error'
-				})
-			}
-		} else if (typeof req.query.question == 'undefined') {
-			logger.log('verbose', `[get /student] user=(${JSON.stringify(user)}) myRes = (cD[req.session.class].students[req.session.username].pollRes.buttonRes) myTextRes = (cD[req.session.class].students[req.session.username].pollRes.textRes) lesson = (cD[req.session.class].lesson)`)
-
-			res.render('pages/student', {
-				title: 'Student',
-				user: JSON.stringify(user),
-				myRes: classInformation[req.session.class].students[req.session.username].pollRes.buttonRes,
-				myTextRes: classInformation[req.session.class].students[req.session.username].pollRes.textRes,
-				lesson: classInformation[req.session.class].lesson
-			})
-		}
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
-/* This is for when you send poll data via a post command or when you submit a quiz.
-If it's a poll it'll save your response to the student object and the database.
-- Riley R., May 24, 2023
-*/
-app.post('/student', isAuthenticated, permCheck, (req, res) => {
-	try {
-		logger.log('info', `[post /student] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
-		logger.log('verbose', `[post /student] poll=(${JSON.stringify(req.query.poll)}) question=(${JSON.stringify(req.body.question)})`)
-
-		if (req.query.poll) {
-			let answer = req.body.poll
-			if (answer) {
-				classInformation[req.session.class].students[req.session.username].pollRes.buttonRes = answer
-			}
-			res.redirect('/poll')
-		}
-		if (req.body.question) {
-			let results = req.body.question
-			let totalScore = 0
-			for (let i = 0; i < classInformation[req.session.class].quiz.questions.length; i++) {
-				if (results[i] == classInformation[req.session.class].quiz.questions[i][1]) {
-					totalScore += classInformation[req.session.class].quiz.pointsPerQuestion
-				} else {
-					continue
-				}
-			}
-			classInformation[req.session.class].students[req.session.username].quizScore = Math.floor(totalScore) + '/' + classInformation[req.session.class].quiz.totalScore
-
-
-			let user = structuredClone(classInformation[req.session.class].students[req.session.username])
-			delete user.API
-			logger.log('verbose', `[post /student] user=(${JSON.stringify(user)}) totalScore=(${totalScore})`)
-
-			res.render('pages/results', {
-				totalScore: Math.floor(totalScore),
-				maxScore: classInformation[req.session.class].quiz.totalScore,
-				title: 'Results'
-			})
-		}
-	} catch (err) {
-		logger.log('error', err.stack);
-		res.render('pages/message', {
-			message: `Error Number ${logNumbers.error}: There was a server error try again.`,
-			title: 'Error'
-		})
-	}
-})
-
 // 404
 app.use((req, res, next) => {
 	try {
@@ -1451,6 +676,9 @@ app.use((req, res, next) => {
 		})
 	}
 })
+
+// Initialize websocket routes
+initSocketRoutes(io);
 
 http.listen(420, async () => {
 	whitelistedIps = await getIpAccess('whitelist')
