+++ routes/apikey.js
@@ -1,5 +1,3 @@
-// The page displaying the API key used when handling oauth2 requests from outside programs such as formPix
-
 const { isLoggedIn } = require("../modules/authentication")
 const { classInformation } = require("../modules/class")
 const { logNumbers } = require("../modules/config")
@@ -7,6 +5,7 @@ const { logger } = require("../modules/logger")
 
 module.exports = {
     run(app) {
+        // The page displaying the API key used when handling oauth2 requests from outside programs such as formPix
         app.get('/apikey', isLoggedIn, (req, res) => {
             try {
                 logger.log('info', `[get /apikey] ip=(${req.ip}) session=(${JSON.stringify(req.session)})`)
